requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
protobuf==3.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.24.2
scipy==1.7.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.1
urllib3==1.26.6
Werkzeug==2.0.1
